# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transfer Funds  API",
        "contact": {
            "email": "r.liuborets@andersenlab.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/transactions": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create transaction",
                "description": "Create transaction for transfer funds",
                "operationId": "createTransaction",
                "requestBody": {
                    "description": "Created transaction",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TransactionData": {
                "properties": {
                    "senderWalletHash": {
                        "type": "string"
                    },
                    "receiverWalletHash": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "commissionPayer": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TransactionData"
                }
            }
        }
    },
    "security": [
        []
    ]
}